♻️ Extract the core logic about subcommand line *pull* into new sub-package *component*.
