+++ ci/surveillance/component/pull.py
@@ -6,7 +6,7 @@ from fake_api_server._utils import YAML
 from fake_api_server.model.api_config import FakeAPIConfig
 from fake_api_server.model.api_config.template._divide import DivideStrategy
 
-from .model.subcmd_pull import PullApiDocConfigArgs
+from ci.surveillance.model.subcmd_pull import PullApiDocConfigArgs
 
 logger = logging.getLogger(__name__)
